backend/requirements.txt
# Django Core
Django==5.2.1
djangorestframework==3.16.0

# Authentication
djangorestframework-simplejwt

# CORS Support
django-cors-headers

# API Documentation
drf-yasg==1.21.10

# PDF Generation
reportlab

# Data Visualization
matplotlib

# Other dependencies
asgiref==3.8.1
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
